refactor: move race test prog to go file
